+++ ICD.Connect.Misc.CrestronPro/Devices/OccupancySensors/AbstractCresnetOccupancySensorAdapter.cs
@@ -30,7 +30,9 @@ namespace ICD.Connect.Misc.CrestronPro.Devices.OccupancySensors
 	{
 		#region Events
 
+		public event EventHandler<GenericEventArgs<eOccupancyFeatures>> OnSupportedFeaturesChanged;
 		public event EventHandler<GenericEventArgs<eOccupancyState>> OnOccupancyStateChanged;
+		public event EventHandler<IntEventArgs> OnPeopleCountChanged;
 
 		#endregion
 
@@ -42,12 +44,16 @@ namespace ICD.Connect.Misc.CrestronPro.Devices.OccupancySensors
 		private eOccupancyState m_OccupancyState;
 
 		private readonly CresnetInfo m_CresnetInfo;
+		private eOccupancyFeatures m_SupportedFeatures;
+		private int m_PeopleCount;
 
 		#endregion
 
 		#region Properties
 
 		public CresnetInfo CresnetInfo { get { return m_CresnetInfo; } }
+
+		public eOccupancyFeatures SupportedFeatures { get { return eOccupancyFeatures.Occupancy; } }
 
 		public eOccupancyState OccupancyState
 		{
@@ -62,6 +68,8 @@ namespace ICD.Connect.Misc.CrestronPro.Devices.OccupancySensors
 				OnOccupancyStateChanged.Raise(this, new GenericEventArgs<eOccupancyState>(m_OccupancyState));
 			}
 		}
+
+		public int PeopleCount { get { return 0; } }
 
 		#endregion